+++ Assets/Scripts/Battle/Slimes.cs
@@ -78,9 +78,9 @@ public class Slimes : MonoBehaviour
             FindAnyObjectByType<AudioManager>().Play("hit1");
 
             yield return new WaitForSeconds(0.15f);
-            if(_boss.TakeDamage((10 /*Base Damage*/ + animal.Level * 5) *
+            if(_boss.TakeDamage((int)((10 /*Base Damage*/ + animal.Level * 5) *
                 (((int)animal.Element == (int)_boss.Element || (int)animal.Element == 0) ? 1 :
-                ((int)animal.Element % 3 + 1 == (int)_boss.Element) ? 2 : 0.5f)))
+                ((int)animal.Element % 3 + 1 == (int)_boss.Element) ? 2 : 0.5f))))
             {
                 break;
             }
